Tidy up album location
